requirements.txt
certifi==2020.6.20
chardet==4.0.0
colorzero==1.1
distro==1.5.0
gpiozero==1.6.2
idna==2.10
python-apt==2.2.1
requests==2.25.1
RPi.GPIO==0.7.0
six==1.16.0
spidev==3.5
ssh-import-id==5.10
urllib3==1.26.5
